Move GPT.js to the AI folder
